+++ CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replaced analysisUtils MRS calculation with bespoke version, that accounts properly for concatenated observations (#393).
 - Updated default values for setting automatic multiscales (#393).
 - Use spectral-cube common beam calculations for common beams (#385).
+- Mosaic weights are now generated after convolution and regridding, and will undo any primary beam corrections (#396).
 
 ### Fixed